refactor: migrate smtc pure functions to party-display-core
