+++ app/src-tauri/party-display-core/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod dlna_proxy;
 pub mod system;
+pub mod smtc;
